+ correct get "all the packages" command
Using `go get github.com/indrainel/go-sra-schema-1.5` will yield a "no buildable Go source files" error message when running the command.  This makes sense because there are no go file in the top level package.

The `-d` option to `go get`instructs get to stop after downloading the packages; that is, it instructs get not to install the packages.  The  ("...") finds all packages below that directory